// Web/RMA-Web-Shell/projects/claimcare/src/app/claim-manager/shared/entities/claim-referral-query-type.ts
export class ClaimReferralQueryType {
    referralQueryTypeId: number;
    name: string;
    isActive: boolean;
    departmentId: number;
    priorityId: number;
    firstLineSupportRoleId: number;
    createdBy: string;
    createdDate: Date;
    modifiedBy: string;
    modifiedDate: Date;
    wizardConfigurationId: number;
}